Extract shared CSV feature helpers
