+++ src/components/deriveRegions.js
@@ -1,6 +1,7 @@
 import { isValidLat, isValidLon, parseFlexibleFloat } from "./geoColumns";
 import { getRowFeatureType } from "./featureTypes";
 import { isRowVisibleForTimeline } from "./timelineVisibility";
+import { applyFirstNonEmpty, parseOrderValue } from "./csvFeatureValueHelpers";
 
 const DEFAULT_STYLE = {
   color: "#3388ff",
@@ -185,11 +186,6 @@ export function deriveRegionsFromCsv({
   return { polygons, skipped, skippedByTimeline, reason: null };
 }
 
-function parseOrderValue(value) {
-  const n = Number.parseFloat(String(value ?? "").trim());
-  return Number.isFinite(n) ? n : null;
-}
-
 function getOrderKey(item) {
   return item.orderValue ?? item.index;
 }
@@ -226,18 +222,6 @@ function resolveStyle(rows) {
   }
 
   return resolved;
-}
-
-function applyFirstNonEmpty(target, key, value, parser) {
-  if (target[key] != null && target[key] !== "") return;
-
-  const raw = String(value ?? "").trim();
-  if (!raw) return;
-
-  const parsed = parser ? parser(raw) : raw;
-  if (parsed == null || parsed === "") return;
-
-  target[key] = parsed;
 }
 
 function parseNumber(value) {
